Minor fixes to:  - continue killing compiler warnings on gcc & msvc  - start working on function/data pointer mismatches  - documentation tweaks
Fixed ascMalloc.c memory logging problems.  This has not been tested on linux/unix.  If it fails to compile, please revert to the previous version and let me know.

git-svn-id: https://svn.ascend4.org/trunk@97 f15cd148-d149-4907-926e-73cf92b3ee03

+++ base/generic/compiler/fractions.c
@@ -48,7 +48,7 @@ short GCD(register int s1, register int s2)
     s2 = -s2;
   }
   if ((s1==0)||(s2==0)) {
-    return s1+s2;
+    return (short)(s1+s2);
   }
   while (s1!=s2) {
     if (s1>s2) {
@@ -57,7 +57,7 @@ short GCD(register int s1, register int s2)
       s2 -= s1;
     }
   }
-  return s1;
+  return (short)s1;
 }
 
 struct fraction Simplify(struct fraction f)
